requirements.txt
requests~=2.25.1
openai~=0.6.3
jsonlines~=2.0.0
matplotlib~=3.3.4
nltk~=3.5
numpy
schedule
pandas
python-Levenshtein
python-crontab
numba~=0.53.1
six~=1.15.0
top2vec~=1.0.21
top2vec[sentence_encoders]
torch~=1.9.0
transformers~=4.3.3
vadersentiment~=3.3.2
python-dateutil~=2.8.1
lxml~=4.6.2
selectorlib~=0.16.0
tqdm~=4.58.0
